feat: integrate with api and add login API functionality
