+++ src/api/userAPI/userAPI.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 import { METHODS, request } from '../axiosInstances';
-import { SignInParams } from '../../models/apiTypes';
+import { LoginResponse } from '../../models/apiTypes';
 
 export default {
-  signIn: (params: SignInParams) => request(METHODS.POST, '/sign-in/', { params }),
+  signIn: (params: LoginResponse) => request(METHODS.POST, '/sign-in/', { params }),
 };
