+++ src/models/apiTypes.tsx
@@ -1,4 +1,11 @@
-export type SignInParams = {
+export interface LoginFormType {
+  email: string;
+  password: string;
+  rememberMe: boolean;
+}
+
+export interface LoginResponse {
   username: string;
   password: string;
-};
+  token?: string;
+}
